added collision detection for candies and kids

+++ src/main/java/model/kids/Kid.java
@@ -12,9 +12,9 @@ public abstract class Kid extends Entity {
 	protected int hp;
 	protected boolean expired;
 
-	protected double yHead;				// head y position (x always same as body)
-	protected double rHead;				// head radius
-	protected double rBody;				// body radius
+	protected float yHead;				// head y position (x always same as body)
+	protected float rHead;				// head radius
+	protected float rBody;				// body radius
 	
 	protected static final double LEFT_WALL = Model.leftBoundary;
 	protected static final double RIGHT_WALL = Model.rightBoundary;
@@ -38,15 +38,15 @@ public abstract class Kid extends Entity {
 		return ID;
 	}
 	
-	public double getHeadY() {
+	public float getHeadY() {
 		return yHead;
 	}
 	
-	public double getHeadRadius() {
+	public float getHeadRadius() {
 		return rHead;
 	}
 	
-	public double getBodyRadius() {
+	public float getBodyRadius() {
 		return rBody;
 	}
